Rename DomainSearch directory
