+++ packages/domain-search/src/components/domain-search/style.scss
