+++ packages/domain-search/src/components/domains-full-cart-items/test/index.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { DomainsFullCartItems } from '..';
-import { DomainSearchContext } from '../../DomainSearch/DomainSearch';
+import { DomainSearchContext } from '../../domain-search';
 
 const defaultContextValue = {
 	isFullCartOpen: false,
