+++ packages/domain-search/src/components/domains-full-cart/test/index.test.tsx
@@ -4,7 +4,7 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { DomainsFullCart } from '../';
-import { DomainSearchContext } from '../../DomainSearch/DomainSearch';
+import { DomainSearchContext } from '../../domain-search';
 
 const defaultContextValue = {
 	isFullCartOpen: false,
