+++ packages/domain-search/src/index.ts
@@ -1,4 +1,4 @@
-export { DomainSearch, type DomainSearchCart } from './components/DomainSearch/DomainSearch';
+export { DomainSearch, type DomainSearchCart } from './components/domain-search';
 export { DomainSearchControls } from './components/DomainSearchControls/DomainSearchControls';
 export { DomainsMiniCart } from './components/DomainsMiniCart/DomainsMiniCart';
 export { DomainsFullCart } from './components/domains-full-cart';
